% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{acept.table}
\alias{acept.table}
\title{acept.table gera uma tabela de aceitação ou rejeição de uma pauta. Na versão atual, só funciona com as categorias
'progressista' e 'conservador'.}
\usage{
acept.table(dados, var)
}
\arguments{
\item{dados}{tabela de dados}

\item{var}{número da variável na tabela de dados}
}
\description{
acept.table gera uma tabela de aceitação ou rejeição de uma pauta. Na versão atual, só funciona com as categorias
'progressista' e 'conservador'.
}
