% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auxiliar.R
\encoding{utf-8}
\name{wrap.it}
\alias{capitalize}
\alias{rname}
\alias{split}
\alias{to.p}
\alias{trim}
\alias{wrap.it}
\title{Funções auxiliares}
\usage{
wrap.it(x, len = 12)

capitalize(x)

trim(x)

split(x, pattern = "(, )|( e )|/")

to.p(x)

rname(x, wrap = 12, dictionary = "dictionary.txt")
}
\arguments{
\item{x}{Vetor de entrada; character para \code{wrap.it}, \code{trim} e \code{capitalize}, numeric para
\code{to.p}.}

\item{len}{Número máximo de caracteres para cada linha}

\item{pattern}{Padrão regular usado para separar os elementos}
}
\description{
Funções auxiliares para manipulação de textos e números.
}
\details{
A função \code{wrap.it} é usada para gerar os nomes em eixos de gráficos, quebrando a linha em
blocos de no máximo \code{len} caracteres.
A função \code{capitalize} transforma um vetor de texto para Iniciais Maiúsculas.
A função \code{trim} remove espaços no início e final de textos.
A função \code{split} separa cada item de um vetor usando um padrão regular, e devolve um vetor
contendo todos os elementos constituintes.
A função \code{to.p} formata um vetor numérico como porcentagem.
}
\examples{
wrap.it("Texto muito muito extremamente longo e desnecessariamente comprido", 10)
capitalize("texto em minúsculas")
trim("     espaços     ")
split("Um item e outro item, finalmente/no entanto")
}

