% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\encoding{utf-8}
\name{carga}
\alias{carga}
\alias{shinyEnv}
\title{Funções para carregar e tratar dados}
\format{An object of class \code{environment} of length 0.}
\usage{
carga(pattern = "*.csv", remove = TRUE, ...)

shinyEnv
}
\arguments{
\item{pattern}{Padrão de busca para o nome dos arquivos.}

\item{remove}{Lógico. As colunas devem ser tratadas para remover espaços em branco e 1-, 2-, 3-, etc?}

\item{\dots}{Demais argumentos para a função \code{read.csv}.}
}
\description{
Funções padrão para carregar dados a partir de arquivos csv.

Ambiente shiny para geração do relatório
}
\details{
A função \code{carga} carrega todos os arquivos com um determinado padrão (por padrão *.csv)
e devolve um data.frame com todos os dados. Um erro pode ser emitido caso os arquivos não contenham
os mesmos nomes de colunas.
}
\keyword{datasets}

