% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dados.R
\docType{data}
\encoding{utf-8}
\name{carga}
\alias{carga}
\alias{registrar}
\alias{relatorio}
\alias{shinyEnv}
\title{Funções para carregar e tratar dados}
\format{\preformatted{<environment: 0x1dea3398> 
}}
\usage{
carga(pattern = "*.csv", remove = TRUE, ...)

shinyEnv

registrar(idx)

relatorio()
}
\arguments{
\item{pattern}{Padrão de busca para o nome dos arquivos.}

\item{remove}{Lógico. As colunas devem ser tratadas para remover espaços em branco e 1-, 2-, 3-, etc?}

\item{\dots}{Demais argumentos para a função \code{read.csv}.}
}
\description{
Funções padrão para carregar dados a partir de arquivos csv.

Ambiente shiny para geração do relatório

Registra um conjunto de colunas para geração de gráficos e tabelas

Registra um conjunto de colunas para geração de gráficos e tabelas
}
\details{
A função \code{carga} carrega todos os arquivos com um determinado padrão (por padrão *.csv)
e devolve um data.frame com todos os dados. Um erro pode ser emitido caso os arquivos não contenham
os mesmos nomes de colunas.
}
\keyword{datasets}

