% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graf.R
\docType{data}
\encoding{utf-8}
\name{graf}
\alias{graf}
\alias{paleta}
\alias{porc}
\title{Gráficos e tabelas}
\format{\preformatted{ chr [1:9] "#8facbc" "#029db1" "#1f3f68" "#72ba94" "#d4c165" ...
}}
\usage{
graf(dados, idx, col = paleta, main = "", under = FALSE)

porc(dados, idx)

paleta
}
\arguments{
\item{dados}{Tabela de dados gerada pela função \code{\link{carga}}}

\item{idx}{índice das colunas desejadas na variável dados}

\item{col}{Cores a serem utilizadas}

\item{main}{Título do gráfico}

\item{under}{Lógico. O label de cada questão deve ser mostrado sob a barra?}
}
\description{
Gráficos e tabelas padrão.
}
\details{
A função \code{graf} gera um gráfico de barra com as porcentagens de cada nível de um fator.
Essa função deve ser usada apenas em perguntas relacionadas, ou seja, todos os fatores devem ter
exatamente os mesmos níveis.
A função \code{porc} gera uma tabela de porcentagem para ser mostrada juntamente com o gráfico acima.
}
\keyword{datasets}

